requirements.txt
neo4j
google-generativeai
faiss-cpu
sentence-transformers
numpy
requests
wikipedia-api
spacy
pandas
lxml
